+++ src/main/data/data.json
@@ -1,0 +1,7 @@
+[
+  {
+    "id": 1,
+    "name": "aaa",
+    "age": 21
+  }
+]
