--- src/main/scala/objects/utils/DataFrameInitializer.scala
@@ -1,10 +1,0 @@
-package agh.scala.footballanalyzer
-package objects.utils
-
-import org.apache.spark.sql.{DataFrame, SparkSession}
-object DataFrameInitializer {
-  def initDFfromURL(spark: SparkSession, url: String): DataFrame = {
-    import spark.implicits._
-    spark.read.json(Seq(DataLoader.fromURLAsString(url)).toDS())
-  }
-}
